refactor: dto로 분리
